mcp/name_tool/requirements.txt
fastapi
uvicorn
python-dotenv
groq
pydantic
starlette
phonenumbers
requests
email-validator